LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/anchor-download-intercept-reject_currententrychange-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: boo

Harness Error (FAIL), message = Unhandled rejection: boo

PASS event and promise ordering for <a download> intercepted by passing a rejected promise to intercept()
